added style guideline
this is the working style guideline for MagmaDNN. It uses the clang-format program to assert style guidelines.

we are using a slightly modified version of the Google c++ style guideline.

clang-format is easily installed with `sudo apt install clang-format` and VSCode will automatically integrate it with the c/c++ extension.

TODO: .cu lint